layout pages, close #18, close #19
