+++ templates/picture/edit.html.twig
@@ -1,13 +1,9 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Edit Picture{% endblock %}
+{% block title %}Modifier l'image{% endblock %}
 
 {% block body %}
-    <h1>Edit Picture</h1>
+    <h1>Modifier l'image</h1>
 
-    {{ include('picture/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="">back to list</a>
-
-    {{ include('picture/_delete_form.html.twig') }}
+    {{ include('picture/_form.html.twig', {'button_label': 'Modifier'}) }}
 {% endblock %}
