+++ templates/trick/index.html.twig
@@ -12,7 +12,7 @@
     <div class="container">
         <h1>Accueil</h1>
         <hr>
-        <div class="d-flex flex-wrap justify-content-start">
+        <div class="d-flex flex-wrap justify-content-center">
             {% for trick in tricks %}
                 {% set pictureOne = trick.pictures.last() %}
                 <div class="card p-2 m-2 d-flex" style="width:250px">
