+++ templates/video/edit.html.twig
@@ -1,13 +1,9 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Edit Video{% endblock %}
+{% block title %}Modifier la Video{% endblock %}
 
 {% block body %}
-    <h1>Edit Video</h1>
+    <h1>Modifier la Video</h1>
 
-    {{ include('video/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('video_index') }}">back to list</a>
-
-    {{ include('video/_delete_form.html.twig') }}
+    {{ include('video/_form.html.twig', {'button_label': 'Modifier'}) }}
 {% endblock %}
